Refactor handles step 3: move image into a new icon_provider crate
- New `icon_provider` port+adapter: `IconProvider` trait + `HttpIconProvider`
  (download over HTTP + resize). `resize`/`image_extension` become internal; the
  persistence-mixing and dead helpers are dropped (icon storage stays in the data store).
- Repoint consumers: `5_wallet` setup (`fetch_icons`), `iced_ui` NFT view (`fetch_icon`),
  and `handles::store::get` (`resize_standard_dimensions_from_bytes`).
- Delete `handles::image`.
