+++ 01_ports_and_adapters/icon_provider/Cargo.toml
@@ -1,0 +1,9 @@
+[package]
+name = "icon_provider"
+version = "0.1.0"
+edition.workspace = true
+
+[dependencies]
+deps = { path = "../../0_deps" }
+types = { path = "../../1_types" }
+async-trait = { workspace = true }
