+++ 01_ports_and_adapters/icon_provider/src/resize.rs
