+++ 3_handles/Cargo.toml
@@ -13,6 +13,7 @@ deps = { path = "../0_deps" }
 types = { path = "../1_types" }
 store = { path = "../2_store" }
 secrets_store = { path = "../01_ports_and_adapters/secrets_store" }
+icon_provider = { path = "../01_ports_and_adapters/icon_provider" }
 thiserror.workspace = true
