--- 3_handles/src/image/mod.rs
@@ -1,3 +1,0 @@
-pub mod download;
-pub mod image_extension;
-pub mod resize;
