+++ 3_handles/src/lib.rs
@@ -1,6 +1,5 @@
 pub mod app_settings;
 pub mod database_handle;
-pub mod image;
 pub mod statics;
 pub mod store;
 pub mod wallet;
