+++ 5_wallet/Cargo.toml
@@ -14,6 +14,7 @@ store = { path = "../2_store" }
 handles = { path = "../3_handles" }
 ledger_connector = { path = "../01_ports_and_adapters/ledger_connector" }
 secrets_store = { path = "../01_ports_and_adapters/secrets_store" }
+icon_provider = { path = "../01_ports_and_adapters/icon_provider" }
 thiserror = { workspace = true }
 serde = { workspace = true, features = ["derive"] }
