+++ Cargo.toml
@@ -5,6 +5,7 @@ cargo-features = ["profile-rustflags"]
 members = [
     "01_ports_and_adapters/app_path",
     "01_ports_and_adapters/data_stores",
+    "01_ports_and_adapters/icon_provider",
     "01_ports_and_adapters/ledger_connector",
     "01_ports_and_adapters/secrets_store",
     "0_deps",
