+++ iced_ui/Cargo.toml
@@ -20,6 +20,7 @@ deps = { path = "../0_deps" }
 types = { path = "../1_types" }
 store = { path = "../2_store" }
 handles = { path = "../3_handles" }
+icon_provider = { path = "../01_ports_and_adapters/icon_provider" }
 font_and_icons = { path = "../4_font_and_icons" }
 wallet = { path = "../5_wallet" }
 # styles = { path = "styles" }
